+++ DeviceBridgeTests/Controllers/ResyncControllerTests.cs
@@ -11,22 +11,22 @@ namespace DeviceBridge.Controllers.Tests
     public class ResyncControllerTests
     {
         private const string MockDeviceId = "test-device";
-        private Mock<ISubscriptionService> _subscriptionServiceMock;
+        private Mock<ISubscriptionScheduler> _dataSubscriptionServiceMock;
         private ResyncController _resyncController;
 
         [SetUp]
         public void Setup()
         {
-            _subscriptionServiceMock = new Mock<ISubscriptionService>();
-            _resyncController = new ResyncController(LogManager.GetCurrentClassLogger(), _subscriptionServiceMock.Object);
+            _dataSubscriptionServiceMock = new Mock<ISubscriptionScheduler>();
+            _resyncController = new ResyncController(LogManager.GetCurrentClassLogger(), _dataSubscriptionServiceMock.Object);
         }
 
         [Test]
-        [Description("Test to ensure that Resync calls SubscriptionService.SynchronizeDeviceDbAndEngineDataSubscriptionsAsync with correct device ID and parameters.")]
+        [Description("Test to ensure that Resync calls SubscriptionScheduler.SynchronizeDeviceDbAndEngineDataSubscriptionsAsync with correct device ID and parameters.")]
         public void TestRegister()
         {
             _resyncController.Resync(MockDeviceId);
-            _subscriptionServiceMock.Verify(p => p.SynchronizeDeviceDbAndEngineDataSubscriptionsAsync(MockDeviceId, false));
+            _dataSubscriptionServiceMock.Verify(p => p.SynchronizeDeviceDbAndEngineDataSubscriptionsAsync(MockDeviceId, false));
         }
     }
 }
